Made Minute/Second components have an uint32_t parameter and asserts the size is not too large

+++ include/datetime/time/components/Minute.h
@@ -15,7 +15,7 @@ struct Minute : TimeComponent
      *
      * @param minute number of minutes.
      */
-    explicit Minute(uint16_t minute) :
+    explicit Minute(uint32_t minute) :
         TimeComponent(minute) {}
 };
